+++ src/defaults.js
@@ -1,0 +1,18 @@
+export const defaults = {
+  minYear: null, // overrides the value set by maxAge
+  maxYear: 'now',
+  minAge: 0,
+  maxAge: 100,
+  monthFormat: 'short',
+  placeholder: true,
+  defaultDate: null,
+  autoInit: true,
+  leadingZero: true,
+  locale: 'en',
+  selectFuture: false,
+  arrange: 'ymd',
+  yearEl: null,
+  monthEl: null,
+  dayEl: null,
+  roundDownDay: true,
+};